+++ spring-batch-core/src/test/java/org/springframework/batch/core/step/item/FaultTolerantStepFactoryBeanTests.java
@@ -272,7 +272,7 @@ public class FaultTolerantStepFactoryBeanTests {
 
 		factory.setSkipPolicy(new SkipPolicy() {
 			@Override
-			public boolean shouldSkip(Throwable t, int skipCount) throws SkipLimitExceededException {
+			public boolean shouldSkip(Throwable t, long skipCount) throws SkipLimitExceededException {
 				throw new  RuntimeException("Planned exception in SkipPolicy");
 			}
 		});
@@ -300,7 +300,7 @@ public class FaultTolerantStepFactoryBeanTests {
 
 		factory.setSkipPolicy(new SkipPolicy() {
 			@Override
-			public boolean shouldSkip(Throwable t, int skipCount) throws SkipLimitExceededException {
+			public boolean shouldSkip(Throwable t, long skipCount) throws SkipLimitExceededException {
 				throw new  RuntimeException("Planned exception in SkipPolicy");
 			}
 		});
